Three failures in the required Code Idiom & Structural Ratchet Gates job, all
introduced by this branch:

1. Freestanding string .len() ABI — FAIL, and a real defect, not a gate quirk.
   examples/09_embedded/simple_os/arch/arm64/boot/baremetal_stubs.c declared
   `uint32_t len` in RuntimeString while its own comment eight lines above says
   "len MUST be uint64_t — codegen inlines .len() as an i64 load". The two
   _Static_asserts there only pin OFFSETS, and the 4 bytes of padding after a
   32-bit field keep `data` at 16, so both asserts passed while a 64-bit .len()
   load read 4 bytes of length and 4 bytes of garbage. Widened the field and
   the two truncating `(uint32_t)` stores. The x86_64 sibling stub already had
   uint64_t; arm64 was the outlier. Gate now PASS — 8 checks.

2. workspace-root-guard — `run_s2.sh`, a 7-line scratch driver at the workspace
   root, absent from origin/main and referenced by nothing but .gitignore.
   Deleted.

3. check-guard-wiring — 5 NEW unwired guards, all real gates for this lane.
   Wired into a new advisory job in repo-hygiene.yml (they need bootstrap stage
   artifacts this runner lacks, so blocking on them would fail every build).
   Nothing was baselined or opted out. Gate now PASS — 0 NEW unwired.

Two merge conflicts against newer main, both resolved toward main:
native_build_main.spl takes main's extracted native_build_normalize_exit_code
with this lane's worker-spawn trace line kept; server.js drops the pinned
serverVersion "0.1.0" so initializeResult's current 0.2.0 default applies.

+++ examples/09_embedded/simple_os/arch/arm64/boot/baremetal_stubs.c
@@ -96,7 +96,7 @@ typedef struct {
  * i64 load at offset 8. See arch/common/baremetal_runtime.h. */
 typedef struct {
     HeapHeader hdr;
-    uint32_t   len;
+    uint64_t   len;
     char       data[];
 } RuntimeString;
 _Static_assert(offsetof(RuntimeString, len) == 8, "RuntimeString.len must sit at offset 8");
@@ -328,7 +328,7 @@ RuntimeValue rt_string_new(RuntimeValue data, RuntimeValue len_val)
     if (!s) return NIL_VALUE;
     s->hdr.type = HEAP_STRING;
     s->hdr.size = (uint32_t)(sizeof(RuntimeString) + (size_t)len + 1);
-    s->len = (uint32_t)len;
+    s->len = (uint64_t)len;
     const char *src = (const char *)(uintptr_t)data;
     if (src && len > 0) __builtin_memcpy(s->data, src, (size_t)len);
     s->data[len] = '\0';
@@ -343,7 +343,7 @@ RuntimeValue rt_string_from_cstr(const char *cstr)
     if (!s) return NIL_VALUE;
     s->hdr.type = HEAP_STRING;
     s->hdr.size = (uint32_t)(sizeof(RuntimeString) + len + 1);
-    s->len = (uint32_t)len;
+    s->len = (uint64_t)len;
     __builtin_memcpy(s->data, cstr, len);
     s->data[len] = '\0';
     return ENCODE_PTR(s);
